Fix identifier bug.
No longer use cygstart/start. Reasons:

1. The reason why these were used in the first place was to ensure that the DLL error
message pops up...but really how often is that going to happen. Even if it does, the user is probably going to try to launch ugraf.exe manually anyways.

2. The PID of the process returned is the PID of the "start" process. This
is undesirable of course, since we want the PID of ugraf.exe

The way we set the working directory now is just by `chdir`ing to the specified directory. Much
easier. Consequently, cygwin-related util functions have been deleted.

Moreover, with the introduction of nxbin/, it is necessary to try to
launch ugraf.exe from nxbin FIRST. If not, ugii/ugraf.exe essnetially
launches nxbin/ugraf.exe in another process with a different PID, and the
returned PID from the launch command is never to be found.