Fix #16.
A '--nxbin' flag has been provided to attempt to launch the process from
nxbin/. Thus, the default is to always launch from ugii/ugraf.exe. This
has more chance of things like Sketch working. However, this breaks the
identifier.

However, a warning is issued when a process will not be identifiable.